fix: 좋아요 중복 가능, 좋아요를 누르지 않았던 좋아요 취소 가능한 문제 해결

+++ src/main/java/net/binder/api/memberlikebin/service/MemberLikeBinService.java
@@ -3,6 +3,7 @@ package net.binder.api.memberlikebin.service;
 import lombok.RequiredArgsConstructor;
 import net.binder.api.bin.entity.Bin;
 import net.binder.api.bin.service.BinService;
+import net.binder.api.common.exception.BadRequestException;
 import net.binder.api.member.entity.Member;
 import net.binder.api.member.service.MemberService;
 import net.binder.api.memberlikebin.entity.MemberLikeBin;
@@ -21,7 +22,13 @@ public class MemberLikeBinService {
 
     public void saveLike(String email, Long binId) {
         Member member = memberService.findByEmail(email);
+
+        if (memberLikeBinRepository.existsByMember_IdAndBin_Id(member.getId(), binId)) {
+            throw new BadRequestException("이미 좋아요를 누른 쓰레기통입니다.");
+        }
+
         Bin bin = binService.findById(binId);
+
         bin.increaseLike();
         MemberLikeBin memberLikeBin = MemberLikeBin.builder()
                 .member(member)
@@ -31,6 +38,11 @@ public class MemberLikeBinService {
     }
 
     public void deleteLike(String email, Long binId) {
+        Member member = memberService.findByEmail(email);
+
+        if (!memberLikeBinRepository.existsByMember_IdAndBin_Id(member.getId(), binId)) {
+            throw new BadRequestException("좋아요를 누르지 않았던 쓰레기통입니다.");
+        }
         Bin bin = binService.findById(binId);
         bin.decreaseLike();
         memberLikeBinRepository.deleteMemberLikeBinByMember_EmailAndBin_Id(email, binId);
